Revert "docs/ added comments and improved structure"
